Introducing flags and revising !log command
+ introducing flags that will be designed by leading "--"
  to provide extra functionality to commands
  that can be confusing to utilize via normal arguments
+ !log command file is now commented and fairly clear
+ renamed `client` to `bot` (more clear meaning)
+ `bot` and `flags` are new command arguments (for extra functionality)

+++ commands/log.js
@@ -1,74 +1,120 @@
 const Discord = require('discord.js')
 const fs = require('fs')
 
-exports.run = async (message, args) => {
-    let channelName = args[0]
-    if (channelName == undefined) {
-        channelName = message.channel.name
+exports.run = async (bot, message, args, flags) => {
+    console.log('Executing `!log` command!')
+
+    datetimeRegexp = new RegExp([
+        '(20[12][0-9])?\-?([01][0-9])\-([0-3][0-9])',
+        '(T([012][0-9])\:([0-6][0-9])\:?([0-6][0-9])?)?']
+        .join(''))
+
+    hasChannelName = args[0] ?
+        ((args[0].match(datetimeRegexp) === null) === true) : false
+
+    let channelName = hasChannelName ? args[0] : message.channel.name
+    let startString = hasChannelName ? (args[1] || null) : (args[0] || null)
+    let endString = hasChannelName ? (args[2] || null) : (args[1] || null)
+    // minimal command input is !log <startDate/startID>
+
+    /** Date formatting:
+    + Full format: YYYY-MM-DDTHH:MM:SS
+    + Month and day are obligatory
+    + Year and time are optional (if time, it must contain THH:MM); 24H format
+    + No year input yields current year, no time input yields 00:00:00:000
+    */
+
+    let startStringMatches = startString === null ?
+        null : (await startString.match(datetimeRegexp))
+    let endStringMatches = endString === null ?
+        null : (await endString.match(datetimeRegexp))
+    // if any of startString/endString are IDs, matches will return null
+
+    switch (startStringMatches === null) {
+
+        case true: // startString is an ID or null
+            var startID = startString || 
+                Discord.SnowflakeUtil.generate(message.channel.createdAt)
+            break
+
+        case false: // startString is a datetime
+            let startStringMatches = startString.match(datetimeRegexp)
+            // group 0 is full match
+            let startYear = (startStringMatches[1] || String(new Date().getFullYear()))
+            let startMonth = startStringMatches[2]
+            let startDay = startStringMatches[3]
+            // group 4 is time full match (THH:MM:SS)
+            let startHour = startStringMatches[5] || "01" // GMT+1
+            let startMinute = startStringMatches[6] || "00"
+            let startSecond = startStringMatches[7] || "00"
+
+            let startDate = new Date(
+                `${startYear}\-${startMonth}\-${startDay}T${startHour}\:${startMinute}\:${startSecond}`)
+            var startID = Discord.SnowflakeUtil.generate(startDate)
+            break
     }
 
-    let startString = args[1] == undefined ? null : args[1]
-    let endString = args[2] == undefined ? null : args[2]
+    switch (endStringMatches === null) {
 
-    // YYYY-MM-DDTHH:MM:SS
-    // month and day are obligatory
-    // year and time are optional (time must contain THH:MM); 24h format
-    // no year yields current year, no time yields 00:00:00:000
-    datetime_regex = /(20[12][0-9])?\-?([01][0-9])\-([0-3][0-9])(T([012][0-9])\:([0-6][0-9])\:?([0-6][0-9])?)?/
+        case true:
+            var endID = endString || message.id
+            break
 
-    let startStringMatches = startString.match(datetime_regex)
-    let startYear = (startStringMatches[1] || String(new Date().getFullYear()))
-    let startMonth = startStringMatches[2]
-    let startDay = startStringMatches[3]
-    let startHour = startStringMatches[5] || "01" // gmt+1
-    let startMinute = startStringMatches[6] || "00"
-    let startSecond = startStringMatches[7] || "00"
+        case false:
+            let endStringMatches = endString.match(datetimeRegexp)
+            let endYear = (endStringMatches[1] || String(new Date().getFullYear()))
+            let endMonth = endStringMatches[2]
+            let endDay = endStringMatches[3]
+            let endHour = endStringMatches[5] || "01"
+            let endMinute = endStringMatches[6] || "00"
+            let endSecond = endStringMatches[7] || "00"
 
-    let startDate = new Date(`${startYear}\-${startMonth}\-${startDay}T${startHour}\:${startMinute}\:${startSecond}`)
-    let startID = Discord.SnowflakeUtil.generate(startDate)
+            let endDate = new Date(
+                `${endYear}\-${endMonth}\-${endDay}T${endHour}\:${endMinute}\:${endSecond}`)
+            var endID = Discord.SnowflakeUtil.generate(endDate)
+            break
+    }
 
-    let endStringMatches = endString.match(datetime_regex)
-    let endYear = (endStringMatches[1] || String(new Date().getFullYear()))
-    let endMonth = endStringMatches[2]
-    let endDay = endStringMatches[3]
-    let endHour = endStringMatches[5] || "01" // gmt+1
-    let endMinute = endStringMatches[6] || "00"
-    let endSecond = endStringMatches[7] || "00"
-
-    let endDate = new Date(`${endYear}\-${endMonth}\-${endDay}T${endHour}\:${endMinute}\:${endSecond}`)
-    let endID = String(Discord.SnowflakeUtil.generate(endDate))
-
-    // possibly use .get to get ids, and client/bot instead of message...
-    let theChannel = await message.guild.channels.find(
+    let logChannel = await message.guild.channels.find(
         c => c.name == channelName && c.type == 'text')
 
-    let messageLogs = (await theChannel.fetchMessages({ after: startID })
+    let messageLogs = (await logChannel.fetchMessages({ after: startID })
         .then(logs => logs.filter(msg => msg.id <= endID))
         .catch(console.error)
     ).map(msg => ({ [msg.id]: [msg.author.username, msg.content] }))
+    console.log('Fetched something!')
+
 
     let fetchComplete = (logs, end) => {
         return (Object.keys(logs[0])[0] == end)
     }
 
     async function continueFetching(current, nm = 0) {
+
+        // get ID of current logs' last message
         let currentID = Object.keys(current[0])[0]
-        let moreLogs = (await theChannel.fetchMessages({ after: currentID })
+        // continue fetching after current last ID
+        let moreLogs = (await logChannel.fetchMessages({ after: currentID })
             .then(logs => logs.filter(msg => msg.id <= endID))
             .catch(console.error)
         ).map(msg => ({ [msg.id]: [msg.author.username, msg.content] }))
 
         let extendedLogs = moreLogs.concat(current)
-        nnm = Object.keys(extendedLogs).length
-        // need to check if consequent calls for the function do something
-        // because endID may not correspond to a message's id and therefore
-        // fetchComplete will not return true -> possibly change fetchComplete
+
+        /**
+         * We need to check if consequent calls for the function are useful
+         * because endID may not correspond to a message's ID (input =? Date)
+         * so fetchComplete() will not return true if this is the case
+         */
+
+        nnm = Object.keys(extendedLogs).length // new number of messages
         if (!fetchComplete(extendedLogs, endID) && (nnm - nm != 0)) {
             console.log('Continuing fetches...')
             console.log(`Current number of messages: ${nnm}`)
-            // if (Object.keys(extendedLogs).length > 5000) { return extendedLogs }
             return continueFetching(extendedLogs, nnm)
+
         } else {
+            console.log(`Final number of messages: ${nnm}`)
             console.log('Finished fetching!')
             return extendedLogs
         }
@@ -79,23 +125,49 @@ exports.run = async (message, args) => {
         logs = (await continueFetching(logs))
     }
 
+    if (Boolean(flags.get('use-dates')) === true) {
+        for (i = 0; i < logs.length; i++) {
+            for ( [key, msg] of Object.entries(logs[i]) ) {
+                var toDate = Discord.SnowflakeUtil.deconstruct(key)
+                    .date.toLocaleDateString('en-GB',
+                        { 'hour': '2-digit',
+                          'minute': '2-digit',
+                          'second': '2-digit'}
+                    )
+                delete logs[i][key]
+                logs[i][toDate] = msg
+            }
+        }
+    }
     let logFile = await JSON.stringify(logs, null, 2)
 
     await fs.writeFileSync(`${channelName}.json`, logFile, (err) => {
         if (err) throw err
-        console.log('file saved')
+        console.log('File saved!')
     })
 
     let numberOfMessages = Object.keys(logs).length
     let fileStats = await fs.statSync(`${channelName}.json`, (err) => {
         if (err) throw err
     })
-    let fileSize = fileStats.size / 1024.0
+    let fileSize = fileStats.size / 1024.0 // KB
 
     var fileMessage = (await new Discord.RichEmbed()
-        .setTitle(`#${channelName} logfile`)
-        .setDescription(`The file that contains all the conversation logs \
-                since and up to the specified dates in the command call.`)
+        .setTitle(`#${logChannel.name} logs`)
+        .setDescription(`Logfile containing all the message logs in \
+            ${logChannel} between: \
+            \n\`\`\`${Discord.SnowflakeUtil.deconstruct(startID).date
+                .toLocaleDateString('en-GB',
+                    { 'timeZone': 'UTC',
+                      'hour': '2-digit',
+                      'minute': '2-digit',
+                      'second': '2-digit'})}`+
+            `\n${Discord.SnowflakeUtil.deconstruct(endID).date
+                .toLocaleDateString('en-GB',
+                    { 'timeZone': 'UTC',
+                      'hour': '2-digit',
+                      'minute': '2-digit',
+                      'second': '2-digit'})}\`\`\``)
         .setColor('#AA1100')
         .setTimestamp(message.createdAt)
         .addField('Caller: ', `<@${message.author.id}>`, true)
@@ -103,12 +175,24 @@ exports.run = async (message, args) => {
                 ${Number(fileSize).toFixed(1)} KB`, true)
         .attachFile(`${channelName}.json`))
 
-    await message.channel.send(fileMessage) // send on call channel
+    if (Boolean(flags.get('no-send')) !== true) {
+        // send on call channel
+        await message.channel.send(fileMessage)
+        // check flag to send on log channel
+        if (Boolean(flags.get('send-to')) === true) {
+            await logChannel.send(fileMessage)
+        }
+    }
 
-    await fs.unlinkSync(`${channelName}.json`, (err) => {
-        if (err) throw err
-        console.log(err)
-    })
+    if (Boolean(flags.get('no-delete')) !== true) {
+        await fs.unlinkSync(`${channelName}.json`, (err) => {
+            if (err) throw err
+            console.log(err)
+            console.log('File deleted!')
+        })
+    }
+    console.log('Done!')
+    return 0
 }
 
 exports.config = {
